refactor(pkh-flow): move shared context to shared folder
